% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_sparse_s7.R, R/20_scheme_s7.R
\name{coordinates}
\alias{coordinates}
\alias{accessors}
\alias{values}
\alias{values<-}
\alias{bValues}
\alias{gradientDirections}
\alias{shellIndices}
\alias{shellValues}
\title{Value and coordinate accessors}
\usage{
coordinates(x, ...)

values(x, ...)

values(x, ...) <- value

bValues(x, ...)

gradientDirections(x, ...)

shellIndices(x, ...)

shellValues(x, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{Replacement value.}
}
\description{
Value and coordinate accessors
}
