% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_dicom_s7.R
\name{asciiField}
\alias{asciiField}
\title{Retrieve fields from a DICOM ASCII header}
\usage{
asciiField(x, ...)
}
\arguments{
\item{x}{A \code{dicomMetadata} object.}

\item{pattern}{An optional regular expression to match field names.}
}
\description{
Retrieve fields from a DICOM ASCII header
}
