% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_dicom_s7.R
\name{dicomMetadata}
\alias{dicomMetadata}
\title{The dicomMetadata class (S7)}
\usage{
dicomMetadata(source = character(0), tags = (function (.data) 
 {
    
  stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call.
  = FALSE)
 })(), tagOffset = integer(0), dataOffset = integer(0),
  dataLength = integer(0), explicitTypes = logical(0),
  endian = character(0), asciiFields = character(0),
  transferSyntax = character(0))
}
\arguments{
\item{source}{String naming the source file.}

\item{tags}{Data frame of tag information.}

\item{tagOffset}{Starting offset for tags in the file.}

\item{dataOffset}{Starting offset for pixel data in the file.}

\item{dataLength}{Pixel data length.}

\item{explicitTypes}{Logical value indicating whether explicit types are
used in the file.}

\item{endian}{String naming the endianness of the file.}

\item{asciiFields}{Character vector containing the contents of the ASCII
header, if requested and present in the file.}

\item{transferSyntax}{Transfer syntax string, if specified in the file;
otherwise the empty string.}
}
\description{
This S7 class represents DICOM metadata, which typically contains detailed
information about the scan parameters and subject. This is a value-semantics
replacement for the \code{\linkS4class{DicomMetadata}} reference class,
which is now a thin backwards-compatible shell around this class. Unlike
the other classes ported in this release, no reference-semantics bug was
ever observed here (the original class has no mutating methods) - it is
included for consistency with the rest of the new API.
}
