\name{dictionary}
\docType{data}
\alias{dictionary}
\title{DICOM tag dictionary}
\description{
  This data set provides information about DICOM tags, including descriptions and data types. It is primarily for internal use.
}
\usage{dictionary}
\format{A data frame, with rows corresponding to known tags.}
\keyword{datasets}