% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_scheme_s7.R
\name{diffusionScheme}
\alias{diffusionScheme}
\title{The diffusionScheme class (S7)}
\usage{
diffusionScheme(bValues = integer(0), gradientDirections = (function
  (.data) 
 {
     stop(sprintf("S3 class <\%s> doesn't have a constructor",
  class[[1]]), call. = FALSE)
 })(), shellIndices = integer(0),
  shellValues = integer(0))
}
\arguments{
\item{bValues}{A vector of b-values in seconds per square millimetre, one
per volume acquired.}

\item{gradientDirections}{A matrix of gradient directions, one row per
volume acquired.}

\item{shellIndices}{An integer vector giving the shell index associated
with each volume.}

\item{shellValues}{A vector of b-values associated with each shell.}
}
\description{
This S7 class represents a diffusion MRI acquisition scheme. It encapsulates
a series of diffusion-weighted volume acquisitions, each with an associated
diffusion-sensitising gradient direction, in one or more "shells" with a
given weighting ("b-value"). This is a value-semantics replacement for the
\code{\linkS4class{DiffusionScheme}} reference class, which is now a thin
backwards-compatible shell around this class.
}
