% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{embrace}
\alias{embrace}
\title{Combine similar strings into one}
\usage{
embrace(strings)
}
\arguments{
\item{strings}{A vector, which will be coerced to mode \code{character}.}
}
\value{
A single merged string, with the common prefix and suffix as
  attributes.
}
\description{
Merge a vector of strings with a common prefix and/or suffix into one string
with the unique parts in braces, comma-separated.
}
\examples{
embrace(c("image.hdr", "image.img"))

}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18.
\url{http://www.jstatsoft.org/v44/i08/}.
}
\author{
Jon Clayden
}
