% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image_s7.R
\name{imageBinarise}
\alias{imageBinarise}
\title{Binarise an image}
\usage{
imageBinarise(x, ...)
}
\arguments{
\item{x}{An \code{mriImage} object.}
}
\description{
Named \code{imageBinarise()} rather than \code{binarise()} to avoid
colliding with \code{mmand::binarise()} - see \code{\link{imageThreshold}}.
}
