% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image.R
\name{[,MriImage,missing,missing-method}
\alias{[,MriImage,missing,missing-method}
\alias{[,MriImage,ANY,missing-method}
\alias{[,MriImage,missing,ANY-method}
\alias{[,MriImage,ANY,ANY-method}
\alias{[,MriImage,MriImage,missing-method}
\alias{[<-,MriImage,missing,missing-method}
\alias{[<-,MriImage,ANY,missing-method}
\alias{[<-,MriImage,missing,ANY-method}
\alias{[<-,MriImage,ANY,ANY-method}
\alias{[<-,MriImage,MriImage,missing-method}
\title{Indexing methods for the legacy MriImage class}
\usage{
\S4method{[}{MriImage,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,MriImage,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,missing,missing}(x, i, j, ...) <- value

\S4method{[}{MriImage,ANY,missing}(x, i, j, ...) <- value

\S4method{[}{MriImage,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{MriImage,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{MriImage,MriImage,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An \code{MriImage} object.}

\item{i, j, \dots}{Indexing objects.}

\item{drop}{Scalar value: should unitary dimensions be dropped?}

\item{value}{New value(s) for replacement forms.}
}
\value{
A vector, array or \code{\linkS4class{SparseArray}}.
}
\description{
These are the backwards-compatible \code{[}/\code{[<-} methods for the
legacy \code{\linkS4class{MriImage}} shell.
}
