% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{fx}
\alias{fx}
\alias{fxy}
\alias{fxyz}
\alias{fi}
\title{Shorthand anonymous functions}
\usage{
fx(expr)

fxy(expr)

fxyz(expr)

fi(expr)
}
\arguments{
\item{expr}{A (single or compound) expression forming the body of the
function.}
}
\value{
The function constructed.
}
\description{
These functions provide a shorthand route to simple anonymous functions.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18.
\url{https://www.jstatsoft.org/v44/i08/}.
}
\author{
Jon Clayden
}
