% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image_s7.R
\name{mask}
\alias{mask}
\title{Mask an image}
\usage{
mask(x, ...)
}
\arguments{
\item{x}{An \code{mriImage} object.}

\item{maskImage}{Another \code{mriImage}, defining the mask.}
}
\description{
Mask an image
}
