% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image_s7.R
\name{metadata}
\alias{metadata}
\title{A metadata-only view of an image}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{An \code{mriImage} object.}
}
\description{
Returns a version of an \code{\link{mriImage}} with any voxel data
removed, retaining only its metadata.
}
