% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image_s7.R
\name{mriImage}
\alias{mriImage}
\title{The mriImage class (S7)}
\usage{
mriImage(imageDims = integer(0), voxelDims = integer(0),
  voxelDimUnits = character(0), source = character(0),
  origin = integer(0), xform = (function (.data) 
 {
    
  stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call.
  = FALSE)
 })(), reordered = logical(0), tags = list(),
  data = (structure(function (data = NULL, coords = (function (.data) 
 {
   
   stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call.
  = FALSE)
 })(), dims = integer(0)) 
 {
     data
     coords
     dims
    
  new_object(S7_object(), data = data, coords = coords, dims = dims)
 }, name =
  "sparseArray", parent = structure(function () 
 {
     .Call(S7_object_)
 },
  name = "S7_object", properties = list(), abstract = FALSE, constructor =
  function () 
 {
     .Call(S7_object_)
 }, validator = function (self) 
 {
  
    if (!is_S7_type(self)) {
         "Underlying data is corrupt"
     }
 },
  class = c("S7_class", "S7_object")), properties = list(data =
  structure(list(name = "data", class = structure(list(), class = "S7_any"),
  getter = NULL, setter = NULL, validator = NULL, default = NULL), class =
  "S7_property"), coords = structure(list(name = "coords", class =
  structure(list(class = "matrix", constructor = function (.data) 
 {
    
  stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call.
  = FALSE)
 }, validator = NULL), class = "S7_S3_class"), getter = NULL, setter
  = NULL, validator = NULL, default = NULL), class = "S7_property"), dims =
  structure(list(name = "dims", class = structure(list(class = "integer",
  constructor_name = "integer", constructor = function (.data = integer(0)) 

  .data, validator = function (object) 
 {
     if (base_class(object) != name)
  {
         sprintf("Underlying data must be <\%s> not <\%s>", name,
  base_class(object))
     }
 }), class = "S7_base_class"), getter = NULL,
  setter = NULL, validator = NULL, default = NULL), class = "S7_property")),
  abstract = FALSE, constructor = function (data = NULL, coords = (function
  (.data) 
 {
     stop(sprintf("S3 class <\%s> doesn't have a constructor",
  class[[1]]), call. = FALSE)
 })(), dims = integer(0)) 
 {
     data
    
  coords
     dims
     new_object(S7_object(), data = data, coords = coords,
  dims = dims)
 }, validator = function (self) 
 {
     if
  (is.vector(self@data) && length(self@data) != nrow(self@coords)) 
        
  return("data length and number of coordinate rows must agree")
     if
  (nrow(self@coords) > 0 && ncol(self@coords) != length(self@dims)) 
        
  return("coords must have one column per array dimension")
     NULL
 }, class
  = c("S7_class", "S7_object")))())
}
\arguments{
\item{imageDims}{Integer vector of dimensions.}

\item{voxelDims}{Numeric vector of pixel/voxel spacings.}

\item{voxelDimUnits}{Character vector of spatial and/or temporal spacing
units. Millimetres and seconds (i.e., c("mm","s")) are typical.}

\item{source}{String naming the file(s) that the image was read from.}

\item{origin}{Numeric vector giving the spatial coordinate origin.}

\item{xform}{Numeric matrix giving the NIfTI-style xform matrix.}

\item{reordered}{Logical value indicating whether the image has been
reordered.}

\item{tags}{Named list of arbitrary DICOM-style tags.}

\item{data}{Sparse or dense array of data, or \code{NULL}.}
}
\description{
This S7 class represents an MRI image, made up of some voxel data (stored
as a sparse or dense numeric array) and some metadata, such as the file it
was read from, the voxel dimensions, and so on. This is a value-semantics
replacement for the \code{\linkS4class{MriImage}} reference class, which
is now a thin backwards-compatible shell around this class.
}
