\name{MriImageMetadata-class}
\Rdversion{1.1}
\docType{class}
\alias{MriImageMetadata-class}
\alias{MriImageMetadata}

\title{Class "MriImageMetadata"}
\description{
  This class represents MRI image metadata. An object of this class contains various information about an image, such as its image and voxel dimensions, coordinate origin and storage data type. An object of type \code{\link{MriImage}} can be treated as a metadata object, since the latter is embedded in it.
}
\section{Extends}{
Class \code{"\linkS4class{SerialisableObject}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}
\section{Methods}{
  \describe{
    \item{\code{initialize(imagedims, voxdims, voxunit, source, datatype, origin, endian)}:}{Create a new object of this class.}
    \item{\code{getDataType()}:}{Retrieve a list describing the storage data type of the image. See \code{\link{getDataTypeByNiftiCode}} for details.}
    \item{\code{getDimensionality()}:}{Retrieve a numeric vector of length 1 giving the number of dimensions in the image.}
    \item{\code{getDimensions()}:}{Retrieve a numeric vector giving the dimensions of the image in voxels.}
    \item{\code{getEndianness()}:}{Retrieve the endianness of the original or stored image. Either \code{"big"} or \code{"little"}.}
    \item{\code{getFieldOfView()}:}{Retrieve a numeric vector giving the dimensions of the image in spatial units, typically mm.}
    \item{\code{getOrigin()}:}{Retrieve the coordinate origin of the image.}
    \item{\code{getSource()}:}{Retrieve a character vector of length 1 giving the source file for the image. Returns \code{"internal"} if the image was not read from a file.}
    \item{\code{getVoxelDimensions()}:}{Retrieve a numeric vector giving the dimensions of the image voxels in spatial units, typically mm.}
    \item{\code{getVoxelUnit()}:}{Retrieve the spatial unit used by getVoxelDimensions(). The most common value is \code{"mm"}.}
    \item{\code{isInternal()}:}{Retrieve a logical value indicating whether the image was generated by a function, rather than read from a file.}
    \item{\code{setEndianness(newEndian)}:}{Set the storage endianness of the image.}
    \item{\code{setSource(newSource)}:}{Set the source of the image.}
    \item{\code{summarise()}:}{Prints a summary of the metadata.}
  }
}
\author{Jon Clayden}
\seealso{\code{\link{SerialisableObject}}, which this class extends. \code{\link{newMriImageMetadataFromTemplate}} for creating metadata objects.}
\keyword{classes}
