% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_serialise.R
\name{Optional}
\alias{Optional}
\title{Optional types}
\usage{
Optional(className)
}
\arguments{
\item{className}{A string naming an existing (S4) class.}
}
\description{
This function declares an optional type, a union of the named class (which
must already be defined) and \code{NULL}. It is meant for use in packages.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\seealso{
\code{\link[methods]{setClassUnion}}
}
\author{
Jon Clayden
}
