% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_serialise.R
\docType{class}
\name{SerialisableObject-class}
\alias{SerialisableObject-class}
\alias{SerialisableObject}
\title{The SerialisableObject class}
\description{
This reference class extends \code{\linkS4class{TractorObject}} by adding a
function for simple serialisation of the data fields of an object, either to
a list or a file. This is intended to be used for classes whose state can
meaningfully be restored from a list of standard R objects (not including
transient C/C++ pointers, for example). A serialised object may be
deserialised using the \code{\link{deserialiseReferenceObject}} function.
}
\section{Methods}{

\describe{
\item{\code{serialise(file = NULL)}}{Serialise the object to a list or file}
}}

\seealso{
\code{\link{save}}
}
