% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_scheme_s7.R
\name{shellCounts}
\alias{shellCounts}
\alias{nShells}
\alias{nDirections}
\title{Number of shells and directions in a diffusion scheme}
\usage{
nShells(x, ...)

nDirections(x, ...)
}
\arguments{
\item{x}{A \code{diffusionScheme} object.}
}
\description{
Number of shells and directions in a diffusion scheme
}
