% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_sparse_s7.R
\name{sparseArray}
\alias{sparseArray}
\title{The sparseArray class (S7)}
\usage{
sparseArray(data = NULL, coords = (function (.data) 
 {
    
  stop(sprintf("S3 class <\%s> doesn't have a constructor", class[[1]]), call.
  = FALSE)
 })(), dims = integer(0))
}
\arguments{
\item{data}{Vector of nonzero data values.}

\item{coords}{Integer matrix of nonzero \code{data} locations, one per row.}

\item{dims}{Integer vector of dimensions.}
}
\description{
This S7 class represents an array with any number of dimensions, in which
a significant proportion of entries are zero. The coordinates of nonzero
entries are stored along with their values, with all remaining entries
assumed to be zero. This is a value-semantics replacement for the
\code{\linkS4class{SparseArray}} reference class, which is now a thin
backwards-compatible shell around this class.
}
