% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_dicom_s7.R
\name{tagValue}
\alias{tagValue}
\title{Retrieve a DICOM tag value}
\usage{
tagValue(x, ...)
}
\arguments{
\item{x}{A \code{dicomMetadata} object.}

\item{group, element}{Integer group and element numbers of the tag.}
}
\description{
Retrieve a DICOM tag value
}
