% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image_s7.R
\name{threshold}
\alias{threshold}
\title{Threshold an image}
\usage{
threshold(x, ...)
}
\arguments{
\item{x}{An \code{mriImage} object.}

\item{level}{The threshold level.}

\item{defaultValue}{The value to assign to voxels below the threshold.}
}
\description{
Threshold an image
}
