% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_dir.R, R/files.R
\name{newMriImageFromDicomDirectory}
\alias{newMriImageFromDicomDirectory}
\alias{newMriImageFromFile}
\alias{writeMriImageToFile}
\title{Deprecated functions}
\usage{
newMriImageFromDicomDirectory(dicomDir, readDiffusionParams = FALSE,
  untileMosaics = TRUE)

newMriImageFromFile(fileName, fileType = NULL, metadataOnly = FALSE,
  volumes = NULL, sparse = FALSE, mask = NULL, reorder = TRUE)

writeMriImageToFile(image, fileName = NULL, fileType = NA,
  overwrite = TRUE)
}
\arguments{
\item{dicomDir}{Character vector of length one giving the name of a
directory containing DICOM files.}

\item{readDiffusionParams}{Logical value: should diffusion MRI parameters
(b-values and gradient directions) be retrieved from the files if
possible?}

\item{untileMosaics}{Logical value: should Siemens mosaic images be
converted into 3D volumes? This may occasionally be performed in error,
which can be prevented by setting this value to \code{FALSE}.}

\item{fileName}{File names, with or without appropriate extension.}

\item{fileType}{A character vector of length one, giving the file type
required or expected. If this option is missing, the file type used for
writing images will be taken from the \code{tractorFileType} option. See
Details.}

\item{metadataOnly}{Logical value: if \code{TRUE}, only metadata are read
into the object.}

\item{volumes}{An optional integer vector specifying a subset of volumes to
read (generally to save memory). If given, only the requested volumes in
the 4D file will be read.}

\item{sparse}{Logical value: should the image data be stored in a
\code{\linkS4class{SparseArray}} object?}

\item{mask}{An optional \code{\linkS4class{MriImage}} object representing a
mask, outside of which the image to be read should be considered to be
zero. This can be used to save memory when only a small part of a large
image is of interest. Ignored if \code{sparse} is not \code{TRUE}.}

\item{reorder}{Logical value: should the image data be reordered to LAS?
This is recommended in most circumstances.}

\item{image}{An \code{\linkS4class{MriImage}} object.}

\item{overwrite}{Logical value: overwrite an existing image file? For
\code{writeImageFile}, an error will be raised if there is an existing
file and this is set to FALSE.}
}
\description{
These functions are deprecated, generally in favour of more succint
alternatives.
}

