% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_serialise.R
\docType{class}
\name{TractorObject-class}
\alias{TractorObject-class}
\alias{TractorObject}
\title{The TractorObject class}
\description{
This reference class extends the standard \code{\linkS4class{envRefClass}}
class, adding methods for finding all of the field or methods available for
an object. There is also a method for summarising key elements of the object
as a named character vector, which can be suitable overridden by inheriting
classes. The \code{show} method prints this summary as a labelled list.
}
\section{Methods}{

\describe{
\item{\code{fields()}}{Retrieve a list of all field names}

\item{\code{methods()}}{Retrieve a list of all method names}

\item{\code{summarise()}}{Summarise key aspects of the object}
}}

