% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_dicom_s7.R
\name{transferSyntax}
\alias{transferSyntax}
\title{Retrieve the transfer syntax of a DICOM file}
\usage{
transferSyntax(x, ...)
}
\arguments{
\item{x}{A \code{dicomMetadata} object.}
}
\description{
Retrieve the transfer syntax of a DICOM file
}
