\name{Undocumented}
\alias{angleBetweenVectors}
\alias{buildStepVectors}
\alias{createNiftiMetadata}
\alias{cross}
\alias{deserialiseListObject}
\alias{displayGraphic}
\alias{ensureFileSuffix}
\alias{execute}
\alias{expandFileName}
\alias{extractDataFromMriImage}
\alias{getColourScale}
\alias{getDescriptionForDicomTag}
\alias{getParametersForFileType}
\alias{hasNiftiMagicString}
\alias{inherit}
\alias{interpolatePalette}
\alias{interpolatePng}
\alias{locateExecutable}
\alias{maskPixels}
\alias{maximumIntensityProjection}
\alias{newMriImageFromAnalyze}
\alias{newMriImageFromNifti}
\alias{newMriImageMetadataFromAnalyze}
\alias{newMriImageMetadataFromNifti}
\alias{promote}
\alias{readDiffusionParametersFromMetadata}
\alias{resolveVector}
\alias{stripNul}
\alias{superimposePng}
\alias{vectorCrossProduct}
\alias{vectorLength}
\alias{writeMriImageToAnalyze}
\alias{writeMriImageToNifti}
\alias{writePng}
\title{Undocumented functions}
\description{
  This page covers several functions which are included in the package primarily for internal use, and are therefore not fully documented.
}
