% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{deduplicate}
\alias{deduplicate}
\title{Concatenate and deduplicate vectors}
\usage{
deduplicate(...)
}
\arguments{
\item{...}{One or more vectors of any mode, usually named.}
}
\value{
The concatenated and deduplicated vector.
}
\description{
This function returns its arguments, after concatenating them using \code{c}
and then removing elements with duplicate names. The first element with each
name will remain. Unnamed elements are retained.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18.
\url{http://www.jstatsoft.org/v44/i08/}.
}
\author{
Jon Clayden
}
