% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_dicom.R
\docType{class}
\name{DicomMetadata-class}
\alias{DicomMetadata}
\alias{DicomMetadata-class}
\title{The DicomMetadata class}
\description{
This class represents DICOM metadata, which typically contains detailed
information about the scan parameters and subject.
}
\section{Methods}{

\describe{
\item{\code{getTagValue(group, element)}}{Retrieve the value of a given tag, using an appropriate R type. Returns NA if the tag is missing}
}}

