% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_sparse.R, R/10_mri_image.R
\docType{methods}
\name{[,SparseArray,ANY,ANY-method}
\alias{[,SparseArray,ANY,ANY-method}
\alias{[<-,SparseArray,ANY,ANY-method}
\alias{[,MriImage,missing,missing-method}
\alias{[,MriImage,ANY,missing-method}
\alias{[,MriImage,missing,ANY-method}
\alias{[,MriImage,ANY,ANY-method}
\alias{[<-,MriImage,missing,missing-method}
\alias{[<-,MriImage,ANY,missing-method}
\alias{[<-,MriImage,missing,ANY-method}
\alias{[<-,MriImage,ANY,ANY-method}
\title{Indexing methods}
\usage{
\S4method{[}{SparseArray,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{SparseArray,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{MriImage,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{MriImage,missing,missing}(x, i, j, ...) <- value

\S4method{[}{MriImage,ANY,missing}(x, i, j, ...) <- value

\S4method{[}{MriImage,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{MriImage,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An object of the appropriate type.}

\item{i, j, \dots}{Indexing objects.}

\item{drop}{Scalar value: should unitary dimensions be dropped?}

\item{value}{New value(s) for replacement forms.}
}
\value{
A vector, array or \code{\link{SparseArray}}.
}
\description{
Indexing methods for \code{\link{SparseArray}} and \code{\link{MriImage}}
objects. For the latter class, arguments are passed to the equivalents for
\code{array} or \code{\link{SparseArray}}. For \code{\link{SparseArray}},
indexing may be blank, or by numeric vector or matrix.
}
\author{
Jon Clayden
}
