% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mri_image.R
\docType{class}
\name{MriImage-class}
\alias{MriImage}
\alias{MriImage-class}
\title{The MriImage class}
\description{
This class represents an MRI image. An object of this class is made up of
some voxel data, stored as a sparse or dense numeric array, and some
metadata, such as the file it was read from, the voxel dimensions, and so
on. The group generic functions \code{\link{Math}}, \code{\link{Ops}} and
\code{\link{Summary}} are defined for this class, as are methods for
coercing to and from a standard \code{\link{array}}.
}
\section{Methods}{

\describe{
\item{\code{apply(...)}}{Apply a function to the margins of the image}

\item{\code{binarise()}}{Binarise the image by setting nonzero values to one}

\item{\code{fill(value)}}{Fill the image with a particular value}

\item{\code{getDataAtPoint(...)}}{Obtain the value of the image at a particular point}

\item{\code{getMetadata()}}{Obtain a version of the image with any data removed}

\item{\code{getNonzeroIndices(array = TRUE, positiveOnly = FALSE)}}{Find voxels whose values are not zero}

\item{\code{getSlice(dim, loc)}}{Extract data from a slice of the image along one dimension}

\item{\code{getSparseness()}}{Obtain the proportion of zeroes in the image}

\item{\code{getTag(key)}}{Retrieve the value of a tag with the specified key}

\item{\code{map(fun, ..., sparse = NULL)}}{Replace the current data with the result of a function}

\item{\code{mask(maskImage)}}{Mask the image, setting zero voxels in the mask to zero}

\item{\code{setOrigin(newOrigin)}}{Update the origin of the image}

\item{\code{setSource(newSource)}}{Update the source of the image}

\item{\code{setXform(newXform)}}{Update the xform matrix associated with the image}

\item{\code{threshold(level, defaultValue = 0)}}{Threshold the image by setting values below the threshold level to zero}
}}

