% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_serialise.R
\docType{class}
\name{SerialisableObject-class}
\alias{SerialisableObject}
\alias{SerialisableObject-class}
\title{The SerialisableObject class}
\description{
This reference class extends the standard \code{\linkS4class{envRefClass}}
class, adding a function for simple serialisation of the data fields of an
object, and one for finding all of the methods available for an object. A
serialised object may be deserialised using the
\code{\link{deserialiseReferenceObject}} function.
}
\section{Methods}{

\describe{
\item{\code{fields()}}{Retrieve a list of all field names}

\item{\code{serialise(file = NULL)}}{Serialise the object to a list or file}
}}

