% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom_dir.R
\name{sortDicomDirectory}
\alias{sortDicomDirectory}
\title{Sort a directory of DICOM files into series}
\usage{
sortDicomDirectory(directory, deleteOriginals = FALSE, sortOn = "series",
  useSeriesTime = FALSE)
}
\arguments{
\item{directory}{A length-1 character vector giving the directory to search
for DICOM files. Subdirectories will also be searched.}

\item{deleteOriginals}{A single logical value. If \code{TRUE}, then the
source files will be deleted after being copied to their new locations,
making the operation a move rather than a copy. Nothing will be deleted if
the copy fails.}

\item{sortOn}{The string \code{"series"}, \code{"subject"} or \code{"date"},
or any combination in the order desired. This will be the basis of the
sort, which will be nested if more than one type is specified.}

\item{useSeriesTime}{Logical value. If \code{TRUE}, use the series time
rather than number as the identifier. This can be more reliable if the
files are from multiple scan sessions.}
}
\value{
This function is called for its side effect.
}
\description{
This function sorts a directory containing DICOM files into subdirectories
by series number (DICOM tag 0x0020,0x0011) series time (0x0008,0x0031),
subject name (0x0010,0x0010) and/or scan date (0x0008,0x0020). Each unique
identifier, together with its description, will be used as the name for a
new subdirectory of the specified top-level directory, and all relevant
files will be copied into that subdirectory. Duplicate file names are
disambiguated if necessary.
}
\author{
Jon Clayden
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18.
\url{http://www.jstatsoft.org/v44/i08/}.
}
\seealso{
\code{\link{readDicomDirectory}} for reading DICOM files into an
\code{MriImage} object.
}

