\name{Undocumented}
\alias{angleBetweenVectors}
\alias{bootstrapExperiment}
\alias{buildStepVectors}
\alias{createNiftiMetadata}
\alias{createWorkspaceFromYaml}
\alias{cross}
\alias{describeExperiment}
\alias{deserialiseListObject}
\alias{displayGraphic}
\alias{ensureFileSuffix}
\alias{execute}
\alias{expandFileName}
\alias{extractDataFromMriImage}
\alias{getColourScale}
\alias{getDescriptionForDicomTag}
\alias{getParametersForFileType}
\alias{getWithDefault}
\alias{hasNiftiMagicString}
\alias{inherit}
\alias{insertColumnAt}
\alias{insertRowAt}
\alias{interpolatePng}
\alias{locateExecutable}
\alias{maskPixels}
\alias{maximumIntensityProjection}
\alias{nArguments}
\alias{newMriImageFromAnalyze}
\alias{newMriImageFromNifti}
\alias{newMriImageMetadataFromAnalyze}
\alias{newMriImageMetadataFromNifti}
\alias{producesError}
\alias{promote}
\alias{rawToCharQuiet}
\alias{readDiffusionParametersFromMetadata}
\alias{requireArguments}
\alias{resolveVector}
\alias{restrict}
\alias{splitAndConvertString}
\alias{superimposePng}
\alias{vectorCrossProduct}
\alias{vectorLength}
\alias{writeMriImageToAnalyze}
\alias{writeMriImageToNifti}
\alias{writePng}
\alias{writeReportToYaml}
\title{Undocumented functions}
\description{
  This page covers several functions which are included in the package primarily for internal use, and are therefore not fully documented.
}
